moved unit tests to a different file
ref #156
